Changed record entry structure, modified code
I changed the structure of the record entries to support
deletion and record redirection. An entry will now
contain a status indicating the type of entry, and
now has the option to contain a redirect RID.

I updated the read and insert function to support the
new format.

This commit is UNTESTED